move supergraph service tests to their own file
