+++ apollo-router/src/services/supergraph.rs
@@ -25,6 +25,8 @@ use crate::json_ext::Path;
 use crate::Context;
 
 pub(crate) mod service;
+#[cfg(test)]
+mod tests;
 
 pub type BoxService = tower::util::BoxService<Request, Response, BoxError>;
 pub type BoxCloneService = tower::util::BoxCloneService<Request, Response, BoxError>;
